refactor(application): addapt files for tests
