+++ src/application/domain/worker/controller/index.ts
@@ -1,8 +1,8 @@
-import type { IPresenter } from '@shared/controller';
+import type { IPresenter } from '@shared/domain/controller';
 import type { IWorker, IWorkerModel } from '../model';
 
 export interface InputGetWorkerByIdDto {
-  id: number;
+  id: string;
 }
 
 export abstract class WorkerControllerAbs {
